Refactor diag native kernels and add oracles
Cython diag kernels: accept generic arrays, coerce to contiguous np.float64 buffers, and use _mv locals for safe pointer calls; return numpy.float64 for scalar stats. Updated .pyi type hints accordingly. Refactored _diag_tests to import and call native _ckernels.diag functions directly (removed conditional native checks and nopython branching). Moved numba reference implementations into tests/_oracles/diag.py and updated tests to use these oracles. Simplifies HAC dispatch and unifies backend behavior; tests updated for the new oracle locations.
