+++ tests/diag_tests/test_wald_test.py
@@ -4,16 +4,16 @@ import numpy as np
 
 from SymbolicDSGE._diag_tests.distributions import PvalMethod
 from SymbolicDSGE._diag_tests.status import TestStatus
-from SymbolicDSGE._diag_tests.moment_calculation_utils import (
+from _oracles.diag import (
     jit_fill_centered,
     jit_fill_mean_ax0,
+    jit_fill_symmetric_target_vec,
+    jit_symmetric_outer_prod_2dim,
 )
 from SymbolicDSGE._diag_tests.wald_test import (
     ERR_BAD_SHAPE,
     ERR_LINALG,
     OK,
-    jit_fill_symmetric_target_vec,
-    jit_symmetric_outer_prod_2dim,
     jit_wald_stat_from_mean_and_cov,
     wald_covariance_hac,
     wald_mean_hac,
@@ -223,8 +223,8 @@ def test_moment_fill_helpers_write_mean_and_centered_buffers() -> None:
     mean = np.empty(2, dtype=np.float64)
     centered = np.empty_like(x)
 
-    jit_fill_mean_ax0.py_func(x, mean)
-    jit_fill_centered.py_func(x, mean, centered)
+    jit_fill_mean_ax0(x, mean)
+    jit_fill_centered(x, mean, centered)
 
     np.testing.assert_allclose(mean, x.mean(axis=0))
     np.testing.assert_allclose(centered, x - x.mean(axis=0))
@@ -234,7 +234,7 @@ def test_symmetric_outer_product_reports_bad_output_shape() -> None:
     x = np.arange(6.0, dtype=np.float64).reshape(3, 2)
     out = np.empty((3, 2), dtype=np.float64)
 
-    err = jit_symmetric_outer_prod_2dim.py_func(x, out)
+    err = jit_symmetric_outer_prod_2dim(x, out)
 
     assert err == ERR_BAD_SHAPE
 
@@ -243,7 +243,7 @@ def test_symmetric_outer_product_fills_upper_triangle_moments() -> None:
     x = np.array([[1.0, 2.0], [3.0, -1.0]], dtype=np.float64)
     out = np.empty((2, 3), dtype=np.float64)
 
-    err = jit_symmetric_outer_prod_2dim.py_func(x, out)
+    err = jit_symmetric_outer_prod_2dim(x, out)
 
     assert err == OK
     np.testing.assert_allclose(
@@ -256,16 +256,16 @@ def test_symmetric_target_vector_helper_fills_and_rejects_bad_shapes() -> None:
     target = np.array([[1.0, 0.25], [0.25, 2.0]], dtype=np.float64)
     out = np.empty(3, dtype=np.float64)
 
-    err = jit_fill_symmetric_target_vec.py_func(target, out)
+    err = jit_fill_symmetric_target_vec(target, out)
 
     assert err == OK
     np.testing.assert_allclose(out, np.array([1.0, 0.25, 2.0], dtype=np.float64))
 
     bad_target = np.array([[1.0, 0.1], [0.2, 1.0]], dtype=np.float64)
-    assert jit_fill_symmetric_target_vec.py_func(bad_target, out) == ERR_BAD_SHAPE
+    assert jit_fill_symmetric_target_vec(bad_target, out) == ERR_BAD_SHAPE
 
     bad_out = np.empty(2, dtype=np.float64)
-    assert jit_fill_symmetric_target_vec.py_func(target, bad_out) == ERR_BAD_SHAPE
+    assert jit_fill_symmetric_target_vec(target, bad_out) == ERR_BAD_SHAPE
 
 
 def test_wald_public_wrappers_validate_targets_and_report_failures() -> None:
